Include data

+++ scrap_tweets.py
@@ -1,19 +1,22 @@
 import snscrape.modules.twitter as sntwitter
-from datetime import datetime
+from datetime import datetime, date, timedelta
 import os
 import csv
 
 from utils import PLAYERS
 
 
-def scrap(query, file):
-    print(query)
+def scrap(query, file, replace=True):
     items = sntwitter.TwitterSearchScraper(query).get_items()
+
+    if not replace and os.path.exists(file):
+        print('Skip', query)
+        return
 
     try:
         os.remove(file)
     except Exception:
-        print("No file to remove")
+        pass
 
     with open(file, mode='w', newline='') as file:
         csv_writer = csv.writer(file, delimiter=';')
@@ -37,14 +40,27 @@ def scrap(query, file):
             csv_writer.writerow(features)
             i += 1
 
-            if i % 500 == 0:
-                print(i, "tweets downloaded")
-
-            if i >= 10_000:
+            if i >= 1000 - 1:
                 break
 
+    print(query, ": done with", i, "tweets")
+
+
+start_date = date(2021, 9, 28)
+end_date = date(2021, 11, 28)
+delta = end_date - start_date
+
+RETRY = ["Kante"]
 
 for query_key in PLAYERS:
-    keywords = " ".join(PLAYERS[query_key])
-    query = f"{keywords} lang:en since:2021-09-28 until:2021-11-28"
-    scrap(query, f"tweets/{query_key}.csv")
+
+    if not os.path.exists(f"tweets/{query_key}"):
+        os.mkdir(f"tweets/{query_key}")
+
+    if query_key not in RETRY:
+        continue
+
+    keywords = " OR ".join(map(lambda x: f"\"{x}\"", PLAYERS[query_key]))
+    for i in range(delta.days):
+        query = f"{keywords} lang:en since:{start_date + timedelta(days=i)} until:{start_date + timedelta(days=i + 1)}"
+        scrap(query, f"tweets/{query_key}/{i}.csv", False)
